refactor(client): extract fan-out invocation into xclient_broadcast.go
Move Broadcast, Fork, Inform out of xclient.go into xclient_broadcast.go.
Same package, mechanical move.
